update to version 0.1.3
update to version 0.1.3

+++ KillAsset/Assets/KillAsset/Editor/Scripts/Workflow/UI/AssetTreeHelper.cs
@@ -10,7 +10,7 @@ namespace KA
 {
     public class AssetTreeHelper
     {
-        public static Action<string, int> onCollectDependencies;
+        public static Action<string, float> onCollectDependencies;
 
         public static int BuildID { get; internal set; } = 0;
